Add auto-expiration to the cache implementation
The original cache implementation was inspired by the Exposure Notifications cache implementation, which specifically didn't have a requirement for expiring cache entires because public keys were rotated infrequently and the services restart frequently. Since this is a general-purpose TTL cache now, we need expiration.

Introducing expiration means we need a background goroutine to manage expiration. This necessitates the introduction of a `Stop()` function to terminate the cacher when the system is shutting down. It also requires some more careful locking around when sweeping occurs. A cache is no longer valid after it has been stopped, and all calls except `Stop()` will panic.

To minimize lock contention, this implementation uses a linked list to keep track of cache items as they are inserted into the cache. New items are inserted at the end of the list, and the sweep operation reads from the beginning until it reaches a non-expired item. This avoids needing to read every item in the cache map (which is unordered) at the cost of additional memory overhead to maintain the order. Since we already have to allocate onto the heap for the cache, the linked list is primarily pointers instead of values.

Finally, the cache attempts to zero out as many values as possible during clearing and cleanup to maximize the greatest sweep from the next GC run.

Other high-level changes:

-   `Set()` no longer returns an error. It never should have.
-   Internal storage uses `time.Time` instead of an int to account for clock skew.