Add shell auto-completions support for the CLI (#125)
* Add shell auto-completions support for the CLI

This adds support for dynamically generating CLI auto-completions (predictions) for command flags and arguments. The default behavior is to predict "something" for all flag values except booleans, and to predict "nothing" for command arguments. This is a superior experience over the current implementation which has no shell completions at all, but developers can further enhance the experience by customizing the CLI flags and commands to predict context-relevant information.

This uses the posener/complete/v2 library, which cleverly self-execs the binary with a special environment that triggers completions. This means completions can be written natively in Go and are automatically versioned with the application. Completions can be as simple as a static list of strings or complex with upstream network calls to predict real-time information.

For more information, see github.com/posener/complete at the master branch.

* Parse review feedback

* Add note to where envvars come from