Tweak the grammar of DiffErrStrings outputs (#140)
The previous output is slightly problematic because it contains
grammatical ambigity that is only resolved after the first error message is
printed. It's a [Garden-path sentence](https://en.wikipedia.org/wiki/Garden-path_sentence).

When an error message is large, the user sees `expected error wall_of_text.....`. The obvious mental
parsing of this is incorrect: "[we] expected [the] error [to be] wall_of_text". Once the user finds the remainder of the message after the wall of text, they find out that their mental parse tree should actually be
"[we] expected [that the] error [we got, which was] <wall of text>...". The problem is exacerbated for large error messages where the sentence structure is not visible at a glance.