feat: programmatic fallback log level and mode
This adds the function `logging.NewFromEnvWithDefaults()`. This allows
the program to provide a log level and a mode (dev/prod) that will be
used in the case that the _LOG_LEVEL and/or _LOG_MODE env vars aren't
set.

The initial use case is for the abc CLI, where we want to use the "dev"
log mode unless the user specifically requested otherwise by setting an
env var.