workerpool: add stop-on-error functionality
This is a breaking API change, mostly because the "option pattern" with Go doesn't work well with generics. The Option has to carry the type signature (which is fine), but the call site requires specifying the type signature again, and it felt really bad. I'd like:

```go
pool := workerpool.New[string](3, workerpool.WithStopOnError(true))
```

but the Option also needs a type inference:

```go
pool := workerpool.New[string](3, workerpool.WithStopOnError[string](true))
```

And that would need to be repeated for each option...