sets: introduce sets package for common set-like operations
This introduces some library functions that are not present in `exp/slices` and `exp/maps` today, notably set intersection, union, and subtraction operations.

The package exposes both `slice` and `map` variants, but the `map` functions are named `*MapKeys` to emphasize that set operations are performed on the keys as opposed to the values.