cli: add GetEnv and LookupEnv helpers (#164)
This adds new helpers for defining the environment. By default, they use the `os` functions. However, they can be overridden (e.g. in tests) with custom values. This also adds a new function on the Command that returns a FlagSet that is configured to use the `LookupEnv` function on the Command.