cli: expand flag helpers and add AfterParse (#168)
* cli: expand flag helpers and add AfterParse

This adds new helpers to the flag and flagset components which mirror more of the stdlib flag functions including `Arg()` and `Lookup()`. It also adds helpers for `LookupEnv` and `GetEnv`.

In combination, these items enable a powerful new feature: `AfterParseFuncs. As their name implies, these functions are invoked after flags are parsed (but before the `Parse` function returns). This enables post-processing of flag values to support use cases like:

- Defaulting the value of flagA to flagB's value. Today this requires custom logic in `Run`, because setting flagB's default value to flagA's value is resolved at compile time, not parse time. With `AfterParse`, it's safe to set the value of flagB to flagA, and it will happen automatically when `Parse` is called.

- Setting the value of a flag to an argument. Perhaps for backwards-compatibility or other API guarantees, you want to default the value of a flag to a command-line argument. Today this requires custom logic in `Run`, but with `AfterParse`, it's possible to set these values automatically when `Parse` is called.

* Add examples