Support back-and-forth dialogs in Prompt() tests (#175)
* Support back-and-forth dialogs in Prompt() tests

Why? In the abc CLI we want to test a function that does multiple rounds of back-and-forth
dialog with the user. Currently, this isn't possible, because prompts
don't appear in a test environment. Prompt()
doesn't write to stdout when stdout is a *bytes.Buffer (as created by BaseCommand.Pipe()).

How do we do this? We add a type-assertion to check is stdout is a
*io.PipeWriter. If so, we assume that it's appropriate to print a
prompt. Tests that want to see prompt messages can just use an *io.Pipe.

Also, we add TestBaseCommand_Prompt_Dialog, which demonstrates how to
write a kind of "conversational" tests using prompts.

Alternative considered: rather than checking for stdout to be a
*io.PipeWriter, check that it implements some special interface like
`type promptable interface { promptable() }`. This would be more
flexible, but also more verbose and harder to understand.

* lint

* PR feedback

* PR feedback

* PR feedback

* lint