cli: add PromptAll and PromptTo (#178)
As identified in abcxyz/lumberjack#433, sometimes we need to read the entire value of stdin, until EOF. The current implementation of `Prompt` relies on newline characters for termination. The new implementation has a generic `PromptTo` function that leverages a `bufio.SplitFunc` to customize splitting. It also adds a `PromptAll` function that will read everything until EOF.

This PR also refactors the tests a bit to unify all the tests for a particular function under a single test (with everyting else under subtests).