Add custom "runs-on" key to reusable workflows
This provides users more control over the underlying runner image used to when executing a reusable workflow. The defaults are unchanged.

Note: we already had this functionality for the "go-test" reusable workflow, but it was named "os". This PR deprecates "os" as an input, since it really should have been named "runner" or "runs-on". If "os" is supplied, GitHub Actions will print out a deprecation message (according to the docs). We prefer "runs-on" over "os", but both will work until we remove "os" as an input to that workflow.