Fix IntersectMapKeys first-wins bug
There was a bug in the old version. The intended behavior was for the
the map value to come from the *first* map in the list. However, the
actual behavior was that the map value came from either the first map or
from the *smallest* map, depending on the order of the input maps.

Alternative considered: do a more efficient algorithm. This alternative
wasn't chosen because it would be a lot of complexity, and the
improvement would only matter for large datasets, and would probably be
slower for small datasets. The most efficient algorithm I can think of
would be a hash join:

 - Find the intersection of the keys:
   - Sort the input maps in increasing order of size
   - Iterate through the input maps and keep only the keys that appear
     in every map
 - Iterate again through each map
   - For each key that's already known to be in every map:
     - If this key doesn't already appear in the output map:
       - Add this key's value to the output map

I didn't choose this option due to its complexity probably not being
worth it.