Fix IntersectMapKeys first-wins bug (#339)
There was a bug in the old version. The intended behavior was for the
the map value to come from the first map in the list. However, the
actual behavior was that the map value came from either the first map or
from the smallest map, depending on the size order of the input maps.