Fix load-workflow-variables issue where config file cannot be found (#348)
This change removes the `./` prefix from the input filepath before
passing the value onto the following steps. If `./guardian.yml` is
provided, then this will be sanitized to `guardian.yml`.

This fixes an issue with `load-workflow-variables` action caused by
providing a relative path for `inputs.filepath`, e.g. `./guardian.yml`
instead of `guardian.yml`. Passing this value to the `sparse-checkout`
option of the Checkout step caused the action to not match with the
`guardian.yml` file at the root of the repository, resulting in a file
not found error in the `Load Workflow Variables` step.