cli: support "-" and "@" for pipes and files (#365)
This PR adds support for flags to optionally allow the user to specify
special syntaxes that tell the CLI to read the contents from a pipe,
prompt, or file. This behavior is off by default and must be enabled on
a per-flag basis.

- If a flag sets `AllowFromFile` to true, then if the value for the flag
starts with an "@" symbol, the argument is interpreted as a filepath and
the contents are read from disk. This is theoretically possible today
with shellisms (e.g. `-foo=$(cat bar.txt)`), but this provides a
shell-agnostic implementation. If the file does not start with "@", then
the value is read as a string as normal. As a special case where the
user really wanted the value to be the literal "@", "\\@" can be used to
escape the input.

- If a flag sets `AllowFromPrompt` to true, then if the value for the
flag is exactly the string "-", then the flag expects input from stdin
or a tty. If neither are present, then the user is prompted. As a
special case where the user really wanted the literal value "-", "\\-"
can be used to escape the input.

Again, both of these behaviors are off by default and must be enabled on
a per-flag basis. In general, I don't expect many flags will use this
configuration, but it's useful to be consistent where necessary.

Finally, it looks like there are some new linting errors, since the
linter was failing locally for me, so I fixed those up.