Fix multi-approvers.js not found bug (#368)
Fixes #363, multi-approvers.js not found.

When a reusable workflow uses `require` (via the
[github-script](https://github.com/actions/github-script) action) to
load a Javascript file, that file must be downloaded (via `curl`) from
the git repo.

This PR also fixes another (as yet unreported) bug where the members
JSON file would not be loaded correctly. This was caused by the path to
the JSON file being incorrect. The root of this issue was a
misunderstanding of how the _calling_ and _called_ repos are working
together.

### Test plan

1. I've tested similar code in ian-shafer/playground and
ian-shafer/playground-2.
2. After the pull request is merged, test using the workflow from
ian-shafer/playground (or any other GitHub repo).