feat: global log settings (#416)
# Proposal

Add global logging env vars that is used if no custom logging env var is
provided.

NOTE: All existing CLIs should be updated to set the default global env
vars so that users can easily configure logging globally for a CLI
and/or still override with local settings.

# Background

As we scale the usage of abc packages for cli and logging we need to
provide a consistent and reliable way to set logging configuration
across CLIs. This will be useful for googlers wanting to debug a
service/binary without having to know the custom prefix utilized for
every single service/binary. Imagine an SRE that wants to interact with
one of the abc backed services/binaries - they just need a simple way to
configure outputs.