feat: linter for all repos (#426)
Proof of concept linter workflow that we can require on all
repositories.

* Initially I thought we could just use the `on::paths` attribute to
trigger jobs. But this is not sufficient as this will create the
condition that only files matching these paths that have _changed_ in
the PR trigger these workflows. We want to trigger these workflows if
the repo has any file matching these paths. Otherwise we cannot make the
lint jobs required status checks since they will not reliably run on
every PR.
* The init workflow checks for files have been tested against a shallow
checkout of the chromium repo. Its performant (~1 sec to scan)
* We have a duplicate lint-pkg.yml workflow that only runs in this
context. It is generated via a template and if any changes are made to
the lint.yml we fail unless you regenerate. This is tested:
https://github.com/abcxyz/pkg/actions/runs/14365896307/job/40278605726?pr=426
* TODO: Support reading workflow lint configuration from a file in the
root of the repo.
* TODO: We will need to update the required checks again.
* TODO: Create a new .golangci.yml without specific import ordering

WANT_LGTM=all